SimpleChromaQueryTool/requirements.txt
chromadb>=0.5.0
python-dotenv==1.0.0
numpy<2.0.0